wao_zip/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
beautifulsoup4==4.11.1
certifi==2022.5.18.1
cffi==1.15.1
Django @ file:///Users/runner/miniforge3/conda-bld/django_1646775880490/work
django-bootstrap4==22.1
django-debug-toolbar==3.6.0
django-editorjs-fields==0.2.7
django-etc==1.3.1
django-hitcount==1.3.5
django-misaka==0.2.1
django-taggit==3.0.0
houdini.py==0.1.0
misaka==2.1.1
Pillow==9.2.0
psycopg2 @ file:///Users/builder/ci_310/psycopg2_1642535040511/work
pycparser==2.21
Pygments==2.13.0
pytz @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_1caf090c-01b6-4973-adac-d2eb884a674e7zonb9eb/croots/recipe/pytz_1654762641285/work
soupsieve==2.3.2.post1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
